+++ CHANGES.rst
@@ -7,9 +7,19 @@ Changelog
 ------------------
 Contributors to this version: Ludwig Lierhammer (:user:`ludwiglierhammer`) and Joseph Siddons (:user:`jtsiddons`)
 
+New features and enhancements
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+* add ``encoding`` optional argument to ``cdm_reader_mapper.read_mdf`` and ``cdm_reader_mapper.read_data`` which overrides default value set by model schema if set (:issue:`268`, :pull:`273`).
+
 License and Legal
 ^^^^^^^^^^^^^^^^^
 Updated copyright statements in `LICENSE` (:issue:`271`, :pull:`272`).
+
+Bug fixes
+^^^^^^^^^
+
+* ``cdm_reader_mapper.write_data``: fix doubling of output file name (:pull:`273`)
 
 2.1.0 (2025-04-08)
 ------------------